--- ui/components/lists-list/lists.go
@@ -1,211 +1,0 @@
-package listslist
-
-import (
-	"github.com/charmbracelet/bubbles/help"
-	"github.com/charmbracelet/bubbles/key"
-	"github.com/charmbracelet/bubbles/list"
-	tea "github.com/charmbracelet/bubbletea"
-	"github.com/charmbracelet/log"
-	"github.com/prgrs/clickup/pkg/clickup"
-	"github.com/prgrs/clickup/ui/common"
-	listitem "github.com/prgrs/clickup/ui/components/list-item"
-	"github.com/prgrs/clickup/ui/context"
-)
-
-const id = "lists-list"
-
-type Model struct {
-	id     common.Id
-	list   list.Model
-	ctx    *context.UserContext
-	log    *log.Logger
-	lists  []clickup.List
-	keyMap KeyMap
-
-	Selected clickup.List
-}
-
-func (m Model) Id() common.Id {
-	return m.id
-}
-
-func InitialModel(ctx *context.UserContext, logger *log.Logger) Model {
-	l := list.New([]list.Item{},
-		list.NewDefaultDelegate(),
-		0, 0)
-
-	l.KeyMap.Quit.Unbind()
-	l.KeyMap.CursorUp.Unbind()
-	l.KeyMap.CursorDown.Unbind()
-
-	l.SetShowHelp(false)
-	l.Title = "Lists"
-
-	log := common.NewLogger(logger, common.ResourceTypeRegistry.COMPONENT, id)
-
-	return Model{
-		id:       id,
-		list:     l,
-		ctx:      ctx,
-		Selected: clickup.List{},
-		lists:    []clickup.List{},
-		log:      log,
-	}
-}
-
-type KeyMap struct {
-	CursorUp            key.Binding
-	CursorUpAndSelect   key.Binding
-	CursorDown          key.Binding
-	CursorDownAndSelect key.Binding
-	Select              key.Binding
-}
-
-func DefaultKeyMap() KeyMap {
-	return KeyMap{
-		CursorUp: key.NewBinding(
-			key.WithKeys("k", "up"),
-			key.WithHelp("k, up", "up"),
-		),
-		CursorUpAndSelect: key.NewBinding(
-			key.WithKeys("K", "shift+up"),
-			key.WithHelp("K, shift+up", "up and select"),
-		),
-		CursorDown: key.NewBinding(
-			key.WithKeys("j", "down"),
-			key.WithHelp("j, down", "down"),
-		),
-		CursorDownAndSelect: key.NewBinding(
-			key.WithKeys("J", "shift+down"),
-			key.WithHelp("J, down", "down and select"),
-		),
-		Select: key.NewBinding(
-			key.WithKeys("enter"),
-			key.WithHelp("enter", "select"),
-		),
-	}
-}
-
-func (m Model) KeyMap() KeyMap {
-	return m.keyMap
-}
-
-func (m Model) Help() help.KeyMap {
-	return common.NewHelp(
-		func() [][]key.Binding {
-			return append(
-				m.list.FullHelp(),
-				[]key.Binding{
-					m.keyMap.CursorUp,
-					m.keyMap.CursorUpAndSelect,
-					m.keyMap.CursorDown,
-					m.keyMap.CursorDownAndSelect,
-					m.keyMap.Select,
-				},
-			)
-		},
-		func() []key.Binding {
-			return append(
-				m.list.ShortHelp(),
-				m.keyMap.CursorUp,
-				m.keyMap.CursorDown,
-				m.keyMap.Select,
-			)
-		},
-	).With(common.KeyBindingBack)
-}
-
-func (m *Model) syncList(lists []clickup.List) {
-	m.log.Info("Synchronizing list")
-	m.lists = lists
-
-	items := NewListItem(lists)
-
-	m.list.SetItems(items)
-	m.list.Select(0)
-	m.log.Info("List synchronized")
-}
-
-func (m *Model) Update(msg tea.Msg) tea.Cmd {
-	var cmd tea.Cmd
-	var cmds []tea.Cmd
-
-	switch msg := msg.(type) {
-	case tea.KeyMsg:
-		switch {
-		case key.Matches(msg, m.keyMap.Select):
-			if m.list.SelectedItem() == nil {
-				m.log.Info("List is empty")
-				break
-			}
-			selectedList := m.list.SelectedItem().(listitem.Item).Data().(clickup.List)
-			m.log.Info("Selected list", "id", selectedList.Id, "name", selectedList.Name)
-			m.Selected = selectedList
-			return ListChangedCmd(m.Selected.Id)
-
-		case key.Matches(msg, m.keyMap.CursorDown):
-			m.list.CursorDown()
-
-		case key.Matches(msg, m.keyMap.CursorDownAndSelect):
-			m.list.CursorDown()
-			if m.list.SelectedItem() == nil {
-				m.log.Info("List is empty")
-				break
-			}
-			selectedList := m.list.SelectedItem().(listitem.Item).Data().(clickup.List)
-			m.log.Info("Selected list", "id", selectedList.Id, "name", selectedList.Name)
-			m.Selected = selectedList
-			return common.ListPreviewCmd(m.Selected.Id)
-
-		case key.Matches(msg, m.keyMap.CursorUp):
-			m.list.CursorUp()
-
-		case key.Matches(msg, m.keyMap.CursorUpAndSelect):
-			m.list.CursorUp()
-			if m.list.SelectedItem() == nil {
-				m.log.Info("List is empty")
-				break
-			}
-			selectedList := m.list.SelectedItem().(listitem.Item).Data().(clickup.List)
-			m.log.Info("Selected list", "id", selectedList.Id, "name", selectedList.Name)
-			m.Selected = selectedList
-			return common.ListPreviewCmd(m.Selected.Id)
-		}
-	}
-
-	m.list, cmd = m.list.Update(msg)
-	cmds = append(cmds, cmd)
-
-	return tea.Batch(cmds...)
-}
-
-func (m Model) View() string {
-	return m.list.View()
-}
-
-func (m Model) Init() tea.Cmd {
-	m.log.Infof("Initializing...")
-	return nil
-}
-
-func (m *Model) SetSize(size common.Size) {
-	m.list.SetSize(size.Width, size.Height)
-}
-
-func (m *Model) FolderChanged(id string) error {
-	lists, err := m.ctx.Api.Clickup.GetListsFromFolder(id)
-	if err != nil {
-		return err
-	}
-	m.syncList(lists)
-
-	return nil
-}
-
-func NewListItem(items []clickup.List) []list.Item {
-	result := make([]list.Item, len(items))
-	for i, v := range items {
-		result[i] = listitem.NewItem(v.Name, v.Id, v)
-	}
-	return result
-}
