+++ ui/components/table-tasks/component.go
@@ -1,7 +1,6 @@
 package tabletasks
 
 import (
-	"github.com/charmbracelet/bubbles/help"
 	"github.com/charmbracelet/bubbles/key"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
@@ -79,80 +78,8 @@ func (m *Model) setTableSize(s common.Size) {
 		WithPageSize(pageSize)
 }
 
-type KeyMap struct {
-	table.KeyMap
-	Select key.Binding
-}
-
 func (m Model) KeyMap() KeyMap {
 	return m.keyMap
-}
-
-func DefaultKeyMap() KeyMap {
-	km := table.DefaultKeyMap()
-
-	return KeyMap{
-		KeyMap: table.KeyMap{
-			RowDown:         common.KeyBindingWithHelp(km.RowDown, "down"),
-			RowUp:           common.KeyBindingWithHelp(km.RowUp, "up"),
-			RowSelectToggle: common.KeyBindingWithHelp(km.RowSelectToggle, "select"),
-			PageDown:        common.KeyBindingWithHelp(km.PageDown, "next page"),
-			PageUp:          common.KeyBindingWithHelp(km.PageUp, "previous page"),
-			PageFirst:       common.KeyBindingWithHelp(km.PageFirst, "first page"),
-			PageLast:        common.KeyBindingWithHelp(km.PageLast, "last page"),
-			Filter:          common.KeyBindingWithHelp(km.Filter, "filter"),
-			FilterBlur:      common.KeyBindingWithHelp(km.FilterBlur, "filter blur"),
-			FilterClear:     common.KeyBindingWithHelp(km.FilterClear, "filter clear"),
-			ScrollRight:     common.KeyBindingWithHelp(km.ScrollRight, "scroll right"),
-			ScrollLeft:      common.KeyBindingWithHelp(km.ScrollLeft, "scroll left"),
-		},
-		Select: key.NewBinding(
-			key.WithKeys("enter"),
-			key.WithHelp("enter", "select"),
-		),
-	}
-}
-
-func (m Model) Help() help.KeyMap {
-	km := m.keyMap
-
-	return common.NewHelp(
-		func() [][]key.Binding {
-			return [][]key.Binding{
-				{
-					km.RowDown,
-					km.RowUp,
-					km.RowSelectToggle,
-				},
-				{
-					km.PageDown,
-					km.PageUp,
-					km.PageFirst,
-					km.PageLast,
-				},
-				{
-					km.Filter,
-					km.FilterBlur,
-					km.FilterClear,
-				},
-				{
-					km.ScrollRight,
-					km.ScrollLeft,
-					m.keyMap.Select,
-				},
-			}
-		},
-		func() []key.Binding {
-			return []key.Binding{
-				km.RowDown,
-				km.RowUp,
-				km.RowSelectToggle,
-				km.PageDown,
-				km.PageUp,
-				m.keyMap.Select,
-			}
-		},
-	)
 }
 
 func InitialModel(ctx *context.UserContext, logger *log.Logger) Model {
@@ -245,22 +172,14 @@ func (m *Model) getColumnsKey(cols []Column) []string {
 }
 
 func (m *Model) Update(msg tea.Msg) tea.Cmd {
-	var cmd tea.Cmd
-	var cmds []tea.Cmd
+	var (
+		cmd  tea.Cmd
+		cmds []tea.Cmd
+	)
 
 	switch msg := msg.(type) {
 	case tea.KeyMsg:
-		switch {
-		case key.Matches(msg, m.keyMap.Select):
-			index := m.table.GetHighlightedRowIndex()
-			if m.table.TotalRows() == 0 {
-				m.log.Info("Table is empty")
-				break
-			}
-			taskId := m.tasks[index].Id
-			m.log.Infof("Receive enter: %d", index)
-			cmds = append(cmds, TaskSelectedCmd(taskId))
-		}
+		return m.handleKeys(msg)
 	}
 
 	m.table, cmd = m.table.Update(msg)
