+++ ui/components/tasks-sidebar/keys.go
@@ -1,0 +1,13 @@
+package taskssidebar
+
+import "github.com/charmbracelet/bubbles/viewport"
+
+type KeyMap struct {
+	viewport.KeyMap
+}
+
+func DefaultKeyMap() KeyMap {
+	return KeyMap{
+		KeyMap: viewport.DefaultKeyMap(),
+	}
+}
